fix relu6 to python hardtanh
